IBX-8184: Implemented reusable RichText validator
