+++ src/bundle/Resources/public/js/scripts/fieldType/ezrichtext.js
@@ -1,6 +1,9 @@
+import initValidator from '@ibexa-admin-ui/src/bundle/Resources/public/js/scripts/fieldType/validator/richtext-validator';
+
 (function (global, doc, ibexa, React, ReactDOM) {
     const SELECTOR_FIELD = '.ibexa-field-edit--ezrichtext';
     const SELECTOR_INPUT = '.ibexa-data-source__richtext';
+    const SELECTOR_LABEL = '.ibexa-field-edit__label';
     const SELECTOR_ERROR_NODE = '.ibexa-form-error';
     const selectContent = (config) => {
         const udwContainer = document.querySelector('#react-udw');
@@ -26,62 +29,12 @@
 
     ibexa.addConfig('richText.alloyEditor.callbacks.selectContent', selectContent);
 
-    class EzRichTextValidator extends ibexa.BaseFieldValidator {
-        constructor(config) {
-            super(config);
-
-            this.richtextEditor = config.richtextEditor;
-        }
-        /**
-         * Validates the input
-         *
-         * @method validateInput
-         * @param {Event} event
-         * @returns {Object}
-         * @memberof EzRichTextValidator
-         */
-        validateInput(event) {
-            const fieldContainer = event.currentTarget.closest(SELECTOR_FIELD);
-            const isRequired = fieldContainer.classList.contains('ibexa-field-edit--required');
-            const label = fieldContainer.querySelector('.ibexa-field-edit__label').innerHTML;
-            const isEmpty = !this.richtextEditor.getData().length;
-            const isError = isRequired && isEmpty;
-            const result = { isError };
-
-            if (isError) {
-                result.errorMessage = ibexa.errors.emptyField.replace('{fieldName}', label);
-            }
-
-            return result;
-        }
-    }
-
     doc.querySelectorAll(`${SELECTOR_FIELD} ${SELECTOR_INPUT}`).forEach((container) => {
         const richtextEditor = new ibexa.BaseRichText();
 
         richtextEditor.init(container);
 
-        const validator = new EzRichTextValidator({
-            classInvalid: 'is-invalid',
-            fieldContainer: container.closest(SELECTOR_FIELD),
-            richtextEditor,
-            eventsMap: [
-                {
-                    selector: '.ibexa-data-source__input.ibexa-input--textarea',
-                    eventName: 'input',
-                    callback: 'validateInput',
-                    errorNodeSelectors: [SELECTOR_ERROR_NODE],
-                },
-                {
-                    selector: SELECTOR_INPUT,
-                    eventName: 'blur',
-                    callback: 'validateInput',
-                    errorNodeSelectors: [SELECTOR_ERROR_NODE],
-                },
-            ],
-        });
-
-        validator.init();
+        const validator = initValidator(container, SELECTOR_FIELD, SELECTOR_ERROR_NODE, SELECTOR_INPUT, SELECTOR_LABEL, richtextEditor);
 
         ibexa.addConfig('fieldTypeValidators', [validator], true);
     });
